Fix build

+++ libraries/chain/wasm_interface.cpp
@@ -1181,11 +1181,11 @@ class action_api : public context_aware_api {
          return context.get_receiver();
       }
 
-      void set_action_return_value( array_ptr<char> packed_blob, size_t datalen ) {
+      void set_action_return_value( array_ptr<char> packed_blob, uint32_t datalen ) {
          context.action_return_value.assign( packed_blob.value, packed_blob.value + datalen );
       }
 
-      int evm_execute(array_ptr<unsigned char> trx, size_t size) {
+      int evm_execute(array_ptr<unsigned char> trx, uint32_t size) {
          return ::evm_execute(trx.value, size);
       }
 };
